WIP: multi zone support
This creates a new find_all_zones method that creates an RXV for all
zones that could exist in the receiver. This is discovered by probing
the desc.xml.

This provides the ability to poke at the zones independently as if
they were dedicated stereos, which they mostly act like (there is some
coupling, if both are on the same input, like NET_RADIO, they will end
up playing the same stream). However power and volume are independent
for the 2.

Starting point for discussion on how to do this right.